feat(wallet/api): makeLoggingPresence
