+++ packages/wallet/api/test/test-middleware.js
@@ -2,43 +2,10 @@
 import { test } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { E, Far } from '@endo/far';
-import { Remotable } from '@endo/marshal';
-import { HandledPromise } from '@endo/eventual-send'; // TODO: convince tsc this isn't needed
-import { makeImportContext } from '../src/marshal-contexts.js';
-
-/**
- * @param {string} iface
- * @param {{
- *   applyMethod: (target: unknown, method: string | symbol, args: unknown[]) => void,
- *   applyFunction: (target: unknown, args: unknown[]) => void,
- * }} handler
- */
-const makePresence = (iface, handler) => {
-  let it;
-  const hp = new HandledPromise((resolve, reject, resolveWithPresence) => {
-    it = resolveWithPresence(handler);
-  });
-  assert(it);
-  assert(hp);
-  return Remotable(iface, undefined, it);
-};
-
-/**
- * @param {string} iface
- * @param {(parts: unknown[]) => void} log
- */
-const makeLoggingPresence = (iface, log) => {
-  /** @type {any} */ // TODO: solve types puzzle
-  const it = makePresence(iface, {
-    applyMethod: (target, method, args) => {
-      log(harden(['applyMethod', target, method, args]));
-    },
-    applyFunction: (target, args) => {
-      log(harden(['applyFunction', target, args]));
-    },
-  });
-  return it;
-};
+import {
+  makeImportContext,
+  makeLoggingPresence,
+} from '../src/marshal-contexts.js';
 
 test('presences for actions log their work', async t => {
   const msgs = [];
